average.c

+++ average.c
@@ -3,11 +3,11 @@
 #include<stdio.h>
 int main()
 {
-  int n1,n2,n3,n4,n5;
+  float n1,n2,n3,n4,n5;
   float avg;
   printf("Enter five numbers:");
-  scanf("%d%d%d%d%d",&n1,&n2,&n3,&n4,&n5);
+  scanf("%f%f%f%f%f",&n1,&n2,&n3,&n4,&n5);
   avg=(n1+n2+n3+n4+n5)/(5.0);
-  printf("Average of five numbers=%f",avg);
+  printf("Average of five numbers=%lf",avg);
   return 0;
 }
